requirements.txt
Flask==3.0.0
Pillow==10.1.0
piexif==1.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
